Modify container_of to make it independent from size_t type

+++ emodbus/include/emodbus/base/add/container_of.h
@@ -4,7 +4,7 @@
 
 #ifndef container_of
 
-#define my_offsetof(st, m) ((size_t)(&((st *)0)->m))
+#define my_offsetof(st, m) ((unsigned long)(&((st *)0)->m))
 
 #ifdef __cplusplus
